requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.10.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
requests==2.32.3
beautifulsoup4==4.13.5
lxml==5.4.0
PyPDF2==3.0.1